backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-decouple>=3.8
django-filter>=23.0

razorpay==1.4.1
setuptools>=65.0.0
reportlab>=4.0.0
